+++ lua/lspsaga/floaterm.lua
@@ -1,8 +1,8 @@
 local api = vim.api
 local window = require('lspsaga.window')
-local M = {}
+local term = {}
 
-function M.open_float_terminal(command)
+function term:open_float_terminal(command)
   local cmd = command or os.getenv('SHELL')
 
   -- get dimensions
@@ -10,8 +10,8 @@ function M.open_float_terminal(command)
   local height = api.nvim_get_option('lines')
 
   -- calculate our floating window size
-  local win_height = math.ceil(height * 0.8)
-  local win_width = math.ceil(width * 0.8)
+  local win_height = math.ceil(height * 0.7)
+  local win_width = math.ceil(width * 0.7)
 
   -- and its starting position
   local row = math.ceil((height - win_height) * 0.4)
@@ -29,27 +29,33 @@ function M.open_float_terminal(command)
 
   local content_opts = {
     contents = {},
-    filetype = 'Floaterm',
     enter = true,
+    winblend = 0,
   }
-
-  local cb, cw, _, ow = window.open_shadow_float_win(content_opts, opts)
-  vim.fn.termopen(cmd, {
-    on_exit = function(...)
-      -- M.close_float_terminal()
-    end,
-  })
-  api.nvim_command('setlocal nobuflisted')
-  api.nvim_command('startinsert!')
-  api.nvim_buf_set_var(cb, 'float_terminal_win', { cw, ow })
-end
-
-function M.close_float_terminal()
-  local has_var, float_terminal_win = pcall(api.nvim_buf_get_var, 0, 'float_terminal_win')
-  if not has_var then
-    return
+  if self.term_bufnr then
+    content_opts.bufnr = self.term_bufnr
+    if api.nvim_buf_is_valid(self.term_bufnr) then
+      api.nvim_buf_set_option(self.term_bufnr, 'modified', false)
+    end
   end
-  window.nvim_close_valid_window(float_terminal_win)
+
+  self.term_bufnr, self.term_winid, self.shadow_bufnr, self.shadow_winid =
+    window.open_shadow_float_win(content_opts, opts)
+
+  if not self.first_open then
+    self.first_open = true
+    vim.fn.termopen(cmd, {
+      on_exit = function(...) end,
+    })
+  end
+  vim.cmd('startinsert!')
 end
 
-return M
+function term:close_float_terminal()
+  if self.term_winid and api.nvim_win_is_valid(self.term_winid) then
+    api.nvim_win_hide(self.term_winid)
+    api.nvim_win_hide(self.shadow_winid)
+  end
+end
+
+return term